refactor: move explorer fetchers to server actions
